requirements.txt
openai~=1.65.4
rich~=13.9.4
Flask==3.1.0
python-dotenv~=1.0
werkzeug~=3.1.3